setup.py: Make click dependency optional
This makes it easier to install only the python modules, and not the
w1thermsensor command line utility.

If there is only interest in the python modules, do the following:

    pip install w1thermsensor

If the w1thermsensor utility is invoked when the package has been
installed in this way, an informative error message is displayed,
notifying that click needs to be installed.

To ensure that the w1thermsor utility also works, specify additional
dependencies using [CLI], as such:

    pip install w1thermsensor[CLI]

Signed-off-by: Jonatan Pålsson <jonatan.p@gmail.com>